refac: overviewListData to util
