+++ src/Components/PokemonCard/PokemonCard.component.js
@@ -1,4 +1,4 @@
-import { Card, Heading, CardBody, Spinner } from 'grommet'
+import { Card, Heading, CardBody } from 'grommet'
 import { StyledCardHeader } from './PokemonCard.module'
 import PokemonDetailsTabs from '../PokemonDetailsTabs'
 import { colorByType } from '../../constants'
@@ -9,7 +9,7 @@ const PokemonCard = ({pokemonBasicData}) => {
     const capitalizeName = name.charAt(0).toUpperCase() + name.slice(1)
     const sprite = sprites.other['official-artwork'].front_default
 
-   return pokemonBasicData ? (
+   return pokemonBasicData && (
             <Card width='medium' height="large" margin={{bottom: '30px'}}>
                 <StyledCardHeader pad="small" direction="column" background={colorOfFirstType} sprite={sprite} height='medium'>
                         <p style={{ alignSelf: 'flex-start'}}>#{id}</p>
@@ -20,7 +20,6 @@ const PokemonCard = ({pokemonBasicData}) => {
                 </CardBody>
             </Card>
         )
-        : <Spinner />
 }
 
 export default PokemonCard
